+++ src/geometry.hpp
@@ -6,6 +6,6 @@
 static constexpr f32 kFlat = 0.000001f;
 static constexpr f32 kExpand = 1000000.f;
 
-static f64 getYFromX(juce::Path& p, f64 x) {
+static f32 getYFromX(juce::Path& p, f32 x) {
   return p.getPointAlongPath(f32(x), juce::AffineTransform::scale(1, kFlat)).y * kExpand;
 }
