+++ src/ui_bridge.hpp
@@ -6,8 +6,8 @@ namespace atmt {
 
 struct UIBridge {
   std::atomic<bool> controlPlayhead = false;
-  std::atomic<double> playheadPosition = 0;
-  std::atomic<double> bpm = 120;
+  std::atomic<f32> playheadPosition = 0;
+  std::atomic<f32> bpm = 120;
   std::atomic<u32> numerator = 4;
   std::atomic<u32> denominator = 4;
 };
